Agent/requirements.txt
absl-py==2.3.1
accelerate==0.33.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.25.8
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
atproto==0.0.62
attrs==25.3.0
av==13.1.0
beautifulsoup4==4.14.2
bert-score==0.3.13
bitsandbytes==0.41.3
blis==1.2.1
boto3==1.40.74
botocore==1.40.74
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
datasets==1.17.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker-pycreds==0.4.0
duckduckgo_search==6.1.0
einops==0.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
ezMM==0.3.1
fastapi==0.115.14
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.70.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
ipython==8.37.0
isal==1.8.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libipld==3.1.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==3.0.0
markdown_pdf==1.10
markdownify==0.14.1
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.31.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
murmurhash==1.0.13
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
openai==1.109.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.3
orjsonl==1.0.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
pillow==11.3.0
pillow-avif-plugin==1.5.1
pip==25.2
platformdirs==4.4.0
pluggy==1.6.0
preshed==3.0.10
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pyecharts==2.0.9
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.1.4
pyreqwest_impersonate==0.5.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==1.3.1
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rouge_score==0.1.2
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.12.0
sentence-transformers==3.4.1
sentry-sdk==2.39.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
sty==1.0.6
swanlab==0.7.1
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.22.1
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.2.0
tweepy==4.14.0
typer==0.19.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
wandb==0.18.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.14
weasel==0.4.1
websockets==13.1
wheel==0.45.1
wrapt==1.17.3
xopen==2.0.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
zlib-ng==1.0.0